# Reduce deflate table sizes