Fixes #223

Before:

```

BenchmarkCompressAllocations/level(-2)/flate-12   	    9255	    159487 ns/op	  342272 B/op	      11 allocs/op
BenchmarkCompressAllocations/level(-2)/gzip-12    	   12951	     99418 ns/op	  342448 B/op	      12 allocs/op
BenchmarkCompressAllocations/level(-1)/flate-12   	    1582	    681491 ns/op	 3235203 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(-1)/gzip-12    	    1743	    618542 ns/op	 3235381 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(0)/flate-12    	   16572	     90453 ns/op	  339968 B/op	       9 allocs/op
BenchmarkCompressAllocations/level(0)/gzip-12     	   10854	    129652 ns/op	  340144 B/op	      10 allocs/op
BenchmarkCompressAllocations/level(1)/flate-12    	    2708	    471414 ns/op	 2186626 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(1)/gzip-12     	    2798	    749605 ns/op	 2186802 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(2)/flate-12    	    1167	   1099054 ns/op	 3759493 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(2)/gzip-12     	    1710	    666056 ns/op	 3759669 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(3)/flate-12    	    2234	    694652 ns/op	 2710921 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(3)/gzip-12     	    2455	    547213 ns/op	 2711100 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(4)/flate-12    	    2505	    541069 ns/op	 2710925 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(4)/gzip-12     	    2149	    673418 ns/op	 2711098 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(5)/flate-12    	    1808	    649272 ns/op	 3235207 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(5)/gzip-12     	    1522	    719524 ns/op	 3235382 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(6)/flate-12    	    1795	    703980 ns/op	 3235204 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(6)/gzip-12     	    1795	    611215 ns/op	 3235381 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(7)/flate-12    	    5728	    201804 ns/op	 1006980 B/op	      13 allocs/op
BenchmarkCompressAllocations/level(7)/gzip-12     	    5728	    196232 ns/op	 1007156 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(8)/flate-12    	    6015	    209091 ns/op	 1006980 B/op	      13 allocs/op
BenchmarkCompressAllocations/level(8)/gzip-12     	    5468	    205745 ns/op	 1007156 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(9)/flate-12    	    5230	    207289 ns/op	 1006981 B/op	      13 allocs/op
BenchmarkCompressAllocations/level(9)/gzip-12     	    4812	    243742 ns/op	 1007155 B/op	      14 allocs/op
PASS
```

After:
```
BenchmarkCompressAllocations/level(-2)/flate-12   	   15889	     75134 ns/op	  342272 B/op	      11 allocs/op
BenchmarkCompressAllocations/level(-2)/gzip-12    	   15424	     76300 ns/op	  342448 B/op	      12 allocs/op
BenchmarkCompressAllocations/level(-1)/flate-12   	    2673	    378711 ns/op	 2448774 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(-1)/gzip-12    	    3342	    377507 ns/op	 2448949 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(0)/flate-12    	   17437	     76986 ns/op	  339968 B/op	       9 allocs/op
BenchmarkCompressAllocations/level(0)/gzip-12     	   15076	     82031 ns/op	  340144 B/op	      10 allocs/op
BenchmarkCompressAllocations/level(1)/flate-12    	    3382	    377466 ns/op	 1924486 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(1)/gzip-12     	    3436	    387788 ns/op	 1924662 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(2)/flate-12    	    1971	    591518 ns/op	 2710923 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(2)/gzip-12     	    2073	    516709 ns/op	 2711102 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(3)/flate-12    	    3250	    426246 ns/op	 2186626 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(3)/gzip-12     	    3084	    420084 ns/op	 2186802 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(4)/flate-12    	    2733	    390467 ns/op	 2186626 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(4)/gzip-12     	    3165	    400509 ns/op	 2186802 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(5)/flate-12    	    2797	    417904 ns/op	 2448774 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(5)/gzip-12     	    2455	    456214 ns/op	 2448948 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(6)/flate-12    	    2733	    471116 ns/op	 2448773 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(6)/gzip-12     	    2673	    443633 ns/op	 2448949 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(7)/flate-12    	    6015	    198306 ns/op	 1006979 B/op	      13 allocs/op
BenchmarkCompressAllocations/level(7)/gzip-12     	    5728	    188045 ns/op	 1007155 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(8)/flate-12    	    6684	    195617 ns/op	 1006979 B/op	      13 allocs/op
BenchmarkCompressAllocations/level(8)/gzip-12     	    6331	    193922 ns/op	 1007155 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(9)/flate-12    	    6015	    193829 ns/op	 1006980 B/op	      13 allocs/op
BenchmarkCompressAllocations/level(9)/gzip-12     	    5728	    197447 ns/op	 1007155 B/op	      14 allocs/op
```